server/requirement.txt
aniso8601==9.0.1
asn1crypto==0.24.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cryptography==40.0.2
Flask==2.0.3
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==2.6
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==2.0.1
numpy==1.19.5
packaging==21.3
psycopg2-binary==2.9.9
pycparser==2.21
pycrypto==2.6.1
PyJWT==2.4.0
pyparsing==3.0.7
pypng==0.20220715.0
pytz==2024.1
pyxdg==0.25
qrcode==7.4.2
requests==2.27.1
SecretStorage==2.3.1
six==1.11.0
typing-extensions==4.1.1
urllib3==1.26.18
Werkzeug==2.0.3
XlsxWriter==3.2.0
xmltodict==0.13.0
zipp==3.6.0
zope.event==4.6
zope.interface==5.5.2
